+++ src/main/Inmortal.java
@@ -1,0 +1,8 @@
+package main;
+
+public class Inmortal extends Normal{
+	@Override
+	public Estado morir(Serpiente serpiente) {
+		return this;
+	}
+}
