+++ src/main/Muerto.java
@@ -1,0 +1,11 @@
+package main;
+
+public class Muerto implements Estado{
+	public Estado moverse(Serpiente serpiente) {
+		return this;
+	}
+	
+	public Estado morir(Serpiente serpiente) {
+		return this;
+	}
+}
